+++ packages/fiori/src/NotificationListItem.ts
@@ -1,5 +1,5 @@
 import {
-	isSpace, isDelete, isF10Shift, isEnterShift, isUp, isDown,
+	isSpace, isDelete, isF10Shift, isEnterShift,
 } from "@ui5/webcomponents-base/dist/Keys.js";
 import type { UI5CustomEvent } from "@ui5/webcomponents-base";
 import customElement from "@ui5/webcomponents-base/dist/decorators/customElement.js";
@@ -18,7 +18,6 @@ import willShowContent from "@ui5/webcomponents-base/dist/util/willShowContent.j
 import jsxRenderer from "@ui5/webcomponents-base/dist/renderer/JsxRenderer.js";
 import NotificationListItemImportance from "./types/NotificationListItemImportance.js";
 import NotificationListItemBase from "./NotificationListItemBase.js";
-import type NotificationList from "./NotificationList.js";
 
 // Icons
 import iconSysEnter2 from "@ui5/webcomponents-icons/dist/sys-enter-2.js";
@@ -485,43 +484,6 @@ class NotificationListItem extends NotificationListItemBase {
 		if (isF10Shift(e)) {
 			e.preventDefault();
 		}
-
-		this.focusSameItemOnNextRow(e);
-	}
-
-	focusSameItemOnNextRow(e: KeyboardEvent) {
-		const target = e.target as HTMLElement;
-		if (!target || target.hasAttribute("ui5-menu-item")) {
-			return;
-		}
-
-		const isFocusWithin = this.matches(":focus-within");
-		if (!isFocusWithin || (!isUp(e) && !isDown(e))) {
-			return;
-		}
-
-		e.preventDefault();
-		e.stopImmediatePropagation();
-
-		const list = this.closest("[ui5-notification-list]") as NotificationList;
-		if (!list) {
-			return;
-		}
-
-		const navItems = list.getEnabledItems();
-		// @ts-expect-error TOFIX strictEvents
-		const index = navItems.indexOf(this) + (isUp(e) ? -1 : 1);
-		const nextItem = navItems[index] as NotificationListItemBase;
-		if (!nextItem) {
-			return;
-		}
-
-		const sameItemOnNextRow = nextItem.getHeaderDomRef()!.querySelector(`.${target.className}`) as HTMLElement;
-		if (sameItemOnNextRow && sameItemOnNextRow.offsetParent) {
-			sameItemOnNextRow.focus();
-		} else {
-			nextItem.focus();
-		}
 	}
 
 	_onkeyup(e: KeyboardEvent) {
@@ -584,7 +546,6 @@ class NotificationListItem extends NotificationListItemBase {
 		// NotificationListItem will never be assigned to a variable of type ListItemBase
 		// typescipt complains here, if that is the case, the parameter to the _press event handler could be a ListItemBase item,
 		// but this is never the case, all components are used by their class and never assigned to a variable with a type of ListItemBase
-		// @ts-expect-error
 		this.fireDecoratorEvent("_press", { item: this });
 	}