+++ packages/fiori/src/types/SearchMode.ts
@@ -1,0 +1,19 @@
+/**
+ * Search mode options.
+ * @public
+ */
+enum SearchMode {
+	/**
+	 * Search field with default appearance.
+	 * @public
+	 */
+	Default = "Default",
+
+	/**
+	 * Search field with additional scope select.
+	 * @public
+	 */
+	Scoped = "Scoped",
+}
+
+export default SearchMode;